# public/swagger/web/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Bütçe API",
        "version": "1.0"
    },
    "paths": {
...
        "/employeelists/findAll": {
            "get": {
                "tags": [
                    "Employee Lists"
                ],
                "summary": "Find all employees",
                "operationId": "employeeLists_findAll",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CommonFindAllResultVM"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No result found"
                    }
                }
            }
        },
        "/employeelists/findAllByPaging": {
            "get": {
                "tags": [
                    "Employee Lists"
                ],
                "summary": "Find all employeelists by pagination",
                "operationId": "employeeLists_findAllByPaging",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Search string",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "_page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "_perPage",
                        "in": "query",
                        "description": "Per page",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "_sort",
                        "in": "query",
                        "description": "Order items",
                        "required": false,
                        "schema": {
                            "type": "array",
                            "items": {}
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EmployeeListFindAllByPageResultVM"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No result found"
                    }
                }
            }
        },
        "/employeelists/upload": {
            "post": {
                "tags": [
                    "Employee Lists"
                ],
                "summary": "Upload excel file",
                "operationId": "employeeLists_upload",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad request, returns to validation errors"
                    }
                }
            }
        },
        "/employeelists/previewResults": {
            "get": {
                "tags": [
                    "Employee Lists"
                ],
                "summary": "Get excel file to preview results",
                "operationId": "employeeLists_previewResults",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad request, returns to validation errors"
                    }
                }
            }
        },
...
        "/employeelists/create": {
            "post": {
                "tags": [
                    "Employee Lists"
                ],
                "summary": "Create a new excel list",
                "operationId": "employeeLists_create",
                "requestBody": {
                    "description": "Create new employee",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EmployeeListImport"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad request, returns to validation errors"
                    }
                }
            }
        },
        "/employeelists/importStatus": {
            "get": {
                "tags": [
                    "Employee Lists"
                ],
                "summary": "Get excel file import status",
                "operationId": "employeeLists_importStatus",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad request, returns to validation errors"
                    }
                }
            }
        },
        "/employeelists/reset": {
            "delete": {
                "tags": [
                    "Employee Lists"
                ],
                "summary": "Reset all statuses",
                "operationId": "employeeLists_reset",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad request, returns to validation errors"
                    }
                }
            }
        },
        "/employeelists/update/{listId}": {
            "put": {
                "tags": [
                    "Employee Lists"
                ],
                "summary": "Update employee list",
                "operationId": "employeeLists_update",
                "parameters": [
                    {
                        "name": "listId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Update employee",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EmployeeListSave"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad request, returns to validation errors"
                    }
                }
            }
        },
        "/employeelists/delete/{listId}": {
            "delete": {
                "tags": [
                    "Employee Lists"
                ],
                "summary": "Delete employee list",
                "operationId": "employeeLists_delete",
                "parameters": [
                    {
                        "name": "listId",
                        "in": "path",
                        "description": "Employee uuid",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CommonFindAllResultVM": {
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "string"
                                },
                                "name": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
...
            "EmployeeListFindAllByPageResultVM": {
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/EmployeeListItem"
                        }
                    },
                    "page": {
                        "type": "integer"
                    },
                    "perPage": {
                        "type": "integer"
                    },
                    "totalPages": {
                        "type": "integer"
                    },
                    "totalItems": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "EmployeeListImport": {
                "properties": {
                    "yearId": {
                        "$ref": "#/components/schemas/ObjectId"
                    },
                    "listName": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "EmployeeListItem": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "listName": {
                        "type": "string"
                    },
                    "pernetNumber": {
                        "type": "string"
                    },
                    "workplaceName": {
                        "type": "string"
                    },
                    "customerShortName": {
                        "type": "string"
                    },
                    "customerColor": {
                        "type": "string"
                    },
                    "sows": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "name": {
                        "type": "string"
                    },
                    "middleName": {
                        "type": "string"
                    },
                    "surname": {
                        "type": "string"
                    },
                    "secondSurname": {
                        "type": "string"
                    },
                    "fullname": {
                        "type": "string"
                    },
                    "tckn": {
                        "type": "string"
                    },
                    "countryId": {
                        "type": "string"
                    },
                    "passportNo": {
                        "type": "string"
                    },
                    "genderId": {
                        "type": "string"
                    },
                    "bloodTypeId": {
                        "type": "string"
                    },
                    "phone": {
                        "type": "string"
                    },
                    "mobile1": {
                        "type": "string"
                    },
                    "mobile2": {
                        "type": "string"
                    },
                    "email1": {
                        "type": "string"
                    },
                    "email2": {
                        "type": "string"
                    },
                    "address1": {
                        "type": "string"
                    },
                    "address2": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "EmployeeListSave": {
                "properties": {
                    "listId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "listName": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "ObjectId": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}